# perf: improve bindgen performance and add timing outputs