before

```
Parser time: 267.05ms
Context time: 585.64ms
Codegen time: 4698.74ms
Format time: 10.01ms
Total time: 5561.44ms
```

after

```
Parser time: 34.15ms
Context time: 63.38ms
Codegen time: 324.83ms
Format time: 8.48ms
Total time: 430.84ms
```